Initial any-scale drawing

+++ src/SlickCommon/Canvas/Quad.cs
@@ -5,16 +5,12 @@
     /// </summary>
     public class Quad
     {
-        public int X;
-        public int Y;
-        public int Width;
-        public int Height;
+        public double X;
+        public double Y;
+        public double Width;
+        public double Height;
 
         public Quad(int x, int y, int width, int height) { X = x; Y = y; Width = width; Height = height; }
-
-        /// <summary>
-        /// Precision loss from double to int
-        /// </summary>
-        public Quad(double x, double y, double width, double height) { X = (int)x; Y = (int)y; Width = (int)width; Height = (int)height; }
+        public Quad(double x, double y, double width, double height) { X = x; Y = y; Width = width; Height = height; }
     }
 }
